ir+metainterp: move OpInfo enum to majit-ir::op_info
OpInfo is a 4-variant enum (Unknown / IntBound / Ptr / FloatConst) carrying
`Rc<RefCell<IntBound>>` and `Rc<RefCell<PtrInfo>>` — both type
arguments now live in majit-ir, so the move is a clean lift with no
metainterp dependency. All 7 inherent methods (ptr, int_bound,
is_constant, get_constant_float, get_int_bound, is_nonnull, is_virtual,
get_ptr_info) are pure-leaf and travel with the enum.

metainterp re-exports OpInfo from `majit_ir::op_info` so the public
`crate::optimizeopt::info::OpInfo` path remains unchanged for callers.
